HRRR - Deprecate the HRRR class in favor of data.hrrr.
This deprecates the HRRR class, which was broken up into the data.hrrr
structure in a previous commit. The class was kept around for backwards
compatibility and this change will also include a version bump.
