+++ tests/test_hrrr_preprocessor.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from tests.RME_test_case import RMETestCase
-from weather_forecast_retrieval.hrrr import HRRR
+from weather_forecast_retrieval.data.hrrr import FileLoader
 from weather_forecast_retrieval.hrrr_preprocessor import HRRRPreprocessor
 
 
@@ -52,7 +52,7 @@ class TestHRRRPreprocessor(RMETestCase):
                 'File {} was not written successfully'.format(file)
             )
 
-        metadata, data = HRRR().get_saved_data(
+        metadata, data = FileLoader().get_saved_data(
             pd.to_datetime(self.end_date),
             pd.to_datetime('2018-07-22 03:00'),
             self.BBOX,
@@ -60,7 +60,7 @@ class TestHRRRPreprocessor(RMETestCase):
             output_dir=self.output_path.as_posix(),
             force_zone_number=self.UTM_ZONE_NUMBER)
 
-        self.assertListEqual(
-            list(data.keys()),
+        self.assertCountEqual(
+            data.keys(),
             ['air_temp', 'relative_humidity', 'wind_u', 'wind_v', 'precip_int', 'short_wave']
         )
