+++ weather_forecast_retrieval/__init__.py
@@ -5,11 +5,3 @@
 __author__ = """Scott Havens"""
 __email__ = 'scott.havens@ars.usda.gov'
 __version__ = '0.6.14'
-
-# from . import hrrr, hrrr_archive, rap, utils
-
-from .hrrr import HRRR
-
-__all__ = [
-    HRRR
-]
